feat: ajouter les composants Leaderboard et Radios pour gérer les états de chargement et vides dans SessionScreen
