+++ app/sessionresult/[session_key].tsx
@@ -1,109 +1,14 @@
-import { DriverItem } from '@/components/drivers/DriverItem';
 import Header from '@/components/ui/Header';
 import Layout from '@/components/ui/Layout';
-import { Driver, useFetchDrivers, useFetchPositionBySessionKey, useFetchRadioBySessionKey } from '@/query/hook';
 import { Stack, useLocalSearchParams } from 'expo-router';
 import React, { useState } from 'react';
-import { ActivityIndicator, FlatList, RefreshControl, StyleSheet, View, Text, TouchableOpacity } from 'react-native';
-import { RadioPlayer } from '@/components/radio/RadioPlayer';
-import NoDataFound from '@/components/NoDataFound';
-import { DriverSkeleton } from '@/components/skeleton/DriverSkeleton';
-import { RadioSkeleton } from '@/components/skeleton/RadioSkeleton';
+import { StyleSheet, View, Text, TouchableOpacity } from 'react-native';
+import Leaderboard from '@/components/session/Leaderboard';
+import Radios from '@/components/session/Radios';
 
 const SessionScreen = () => {
     const { session_key } = useLocalSearchParams<{ session_key: string }>();
-    const { data: positions, isRefetching, refetch } = useFetchPositionBySessionKey(session_key);
-    const { data: drivers, isLoading: isDriverLoading } = useFetchDrivers();
-    const { data: radioData, isLoading: isRadioLoading, refetch: refetchRadio } = useFetchRadioBySessionKey(session_key);
-
     const [activeSection, setActiveSection] = useState<'classement' | 'radio'>('classement');
-
-    const classement: Driver[] = (positions ?? []).map((pos) => {
-        const driver = drivers?.find((d) => d.driver_number === pos.driver_number);
-        return { ...driver, ...pos } as Driver;
-    });
-
-    const renderEmptyLeaderboard = () => {
-        if (isDriverLoading) {
-            return <DriverSkeleton />;
-        } else {
-            return <NoDataFound entiyName='leaderboard' />;
-        }
-    }
-
-    const renderEmptyRadio = () => {
-        if (isRadioLoading) {
-            return <RadioSkeleton />;
-        } else {
-            return <NoDataFound entiyName='radio' />;
-        }
-    }
-
-    const renderContent = () => {
-        if (activeSection === 'classement') {
-
-            return (
-                <FlatList
-                    data={classement}
-                    renderItem={({ item, index }) => (
-                        <DriverItem item={item} position={index + 1} />
-                    )}
-                    contentContainerStyle={styles.listContainer}
-                    showsVerticalScrollIndicator={false}
-                    refreshControl={
-                        <RefreshControl
-                            refreshing={isRefetching}
-                            onRefresh={() => {
-                                refetch();
-                                refetchRadio();
-                            }}
-                            tintColor={'#ee0000'}
-                        />
-                    }
-                    windowSize={1}
-                    initialNumToRender={8}
-                    maxToRenderPerBatch={8}
-                    ListEmptyComponent={renderEmptyLeaderboard}
-                />
-            );
-        } else {
-            if (isRadioLoading) {
-                return <ActivityIndicator size="large" color="#ee0000" />;
-            }
-
-            if (radioData && radioData.length > 0) {
-                return (
-                    <FlatList
-                        data={radioData}
-                        renderItem={({ item }) => (
-                            <RadioPlayer radioData={item} />
-                        )}
-                        contentContainerStyle={styles.listContainer}
-                        showsVerticalScrollIndicator={false}
-                        refreshControl={
-                            <RefreshControl
-                                refreshing={false}
-                                onRefresh={() => {
-                                    refetchRadio();
-                                }}
-                                tintColor={'#ee0000'}
-                            />
-                        }
-                        windowSize={1}
-                        initialNumToRender={5}
-                        maxToRenderPerBatch={5}
-                        ListEmptyComponent={renderEmptyRadio}
-                    />
-                );
-            }
-
-            return (
-                <View style={styles.noDataContainer}>
-                    <Text style={styles.noDataText}>Aucun extrait radio disponible</Text>
-                </View>
-            );
-        }
-    };
 
     return (
         <Layout>
@@ -126,7 +31,11 @@ const SessionScreen = () => {
             </View>
 
             <View style={styles.container}>
-                {renderContent()}
+                {activeSection === 'classement' ? (
+                    <Leaderboard session_key={session_key as string} />
+                ) : (
+                    <Radios session_key={session_key as string} />
+                )}
             </View>
         </Layout>
     );
@@ -160,11 +69,6 @@ const styles = StyleSheet.create({
     activeTabText: {
         fontWeight: 'bold',
         color: '#ee0000',
-    },
-    listContainer: {
-        gap: 6,
-        paddingBottom: 20,
-        paddingTop: 10,
     },
     noDataContainer: {
         padding: 20,
